backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==43.0.1
python-multipart==0.0.6
httpx==0.25.2
requests==2.31.0
celery==5.3.4
redis==5.0.1
aiohttp==3.10.11

# New dependencies for recruitment automation pipeline
beautifulsoup4==4.12.2
lxml==4.9.3
pdfminer.six==20221105
python-docx==0.8.11
jinja2==3.1.2
loguru==0.7.0
tenacity==8.2.2
aiosmtplib==1.1.7
email-validator==2.0.0
ratelimit==2.2.1
textract==1.5.0

# Testing and development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
